cli: move the value formatting code to its own files
Release note: None
