+++ pkg/cli/BUILD.bazel
@@ -46,6 +46,7 @@ go_library(
         "nodelocal.go",
         "quit.go",
         "sql.go",
+        "sql_format_value.go",
         "sql_util.go",
         "sqlfmt.go",
         "start.go",
@@ -292,6 +293,7 @@ go_test(
         "log_flags_test.go",
         "main_test.go",
         "nodelocal_test.go",
+        "sql_format_value_test.go",
         "sql_test.go",
         "sql_util_test.go",
         "start_test.go",
